all right. (Some mallocs allocations can be adjusted)

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: ibaby <ibaby@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/31 22:34:24 by ibaby             #+#    #+#             */
-/*   Updated: 2024/05/19 18:56:41 by ibaby            ###   ########.fr       */
+/*   Updated: 2024/05/19 19:34:11 by ibaby            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dst_size)
 {
-	int i;
-	int retval;
+	size_t i;
+	size_t retval;
 
 	i = 0;
 	if (!dst || !src)
@@ -23,7 +23,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dst_size)
 	retval = ft_strlen(src);
 	if (dst_size == 0)
 		return (retval);
-	while (i < (int)dst_size - 1 && src[i])
+	while (i < dst_size - 1 && src[i])
 	{
 		dst[i] = src[i];
 		i++;
